Add x64 build support
Replace inline assembly with __rdtsc() intrinsic, fix GWL_WNDPROC to
GWLP_WNDPROC, fix pointer truncation casts, update SetWindowLong to
SetWindowLongPtr, add x64 platform configurations to vcxproj, build
x64 projectM-eval libs from source, and add Platform parameter to
build.ps1.

+++ Visualizer/vis_milk2/plugin.cpp
@@ -9886,7 +9886,7 @@ char* NextLine(char* p) {
 static unsigned int WINAPI __UpdatePresetList(void* lpVoid) {
   // NOTE - this is run in a separate thread!!!
 
-  DWORD flags = (DWORD)lpVoid;
+  DWORD flags = (DWORD)(uintptr_t)lpVoid;
   bool bForce = (flags & 1) ? true : false;
   bool bTryReselectCurrentPreset = (flags & 2) ? true : false;
